# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VerifiA API
  version: 1.0.0
  description: |
    API REST para el sistema VerifiA de tokens de presencia criptográficos efímeros.

    ## Flujo completo
    1. Portal → `POST /challenges` → obtiene nonce + QR
    2. App escanea QR → realiza FaceTec + App Attest + Passkey
    3. App → `POST /tokens/issue` → recibe JWT ES256
    4. Portal hace polling → `GET /tokens/verify/:nonce`
    5. Portal → `POST /tokens/validate/:nonce` → consume token → muestra ✅/❌

  contact:
    name: Carlos Iván Ibarra Pacheco
    email: A01646019@tec.mx
servers:
  - url: https://api.verifia.dev
    description: Producción (Railway)
  - url: http://localhost:3001
    description: Desarrollo local
...
paths:
  /api/v1/challenges:
    post:
      tags: [Challenges]
      summary: Crear challenge
      description: |
        Crea un nuevo challenge (nonce CSPRNG) para una sesión de verificación.
        El portal llama a este endpoint para generar el QR que el portador escaneará.
        El `verifier_id` se convierte en el claim `aud` del JWT emitido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChallengeRequest'
            example:
              verifier_id: "mi-empresa-api-key-abc123"
      responses:
        '201':
          description: Challenge creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/challenges/incoming/history:
    get:
      tags: [Challenges]
      summary: Historial de solicitudes recibidas (terminales)
      description: |
        Devuelve las solicitudes de verificación dirigidas al correo de la cuenta
        autenticada que ya alcanzaron un estado terminal (USED / EXPIRED / REJECTED /
        CANCELLED). Incluye `receipt_id` cuando se emitió un recibo, para abrir el Ticket.
        Requiere sesión Bearer.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        '200':
          description: Historial de solicitudes recibidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingHistoryResponse'
        '401':
          description: Sesión inválida o ausente
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key del verificador. Define el claim `aud` del JWT.
        Cada organización verificadora recibe una clave única.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión de cuenta (JWT ES256, TTL 7 días). Emitido por
        `/auth/login` y `/auth/set-password`.
  schemas:
    CreateChallengeRequest:
      type: object
      required: [verifier_id]
      properties:
        verifier_id:
          type: string
          minLength: 1
          maxLength: 128
          description: Identificador del verificador (API key)
    ChallengeResponse:
      type: object
      properties:
        nonce:
          type: string
          description: Hex string de 64 chars (32 bytes CSPRNG)
          example: "a3f8c2..."
        expires_in:
          type: integer
          description: Segundos hasta expiración del challenge
          example: 600
        qr_data:
          type: string
          description: Contenido del QR (deep link verifia://)
          example: "verifia://badge?nonce=a3f8c2..."
        deep_link:
          type: string
          description: Deep link para la app móvil
          example: "verifia://badge?nonce=a3f8c2..."
        expires_at:
          type: string
          format: date-time
...
    IncomingHistoryItem:
      type: object
      properties:
        nonce: { type: string }
        status: { type: string }
        verifier_id: { type: string }
        rejection_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        requester:
          type: object
          nullable: true
          properties:
            email: { type: string }
            full_name: { type: string, nullable: true }
            profile_photo: { type: string, nullable: true }
        receipt_id: { type: string, nullable: true }
        verified_at: { type: string, format: date-time, nullable: true }
    IncomingHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IncomingHistoryItem'
        page: { type: integer }
        limit: { type: integer }
...
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
  responses:
    ValidationError:
      description: Payload inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Invalid request body"
            code: "VALIDATION_ERROR"
    RateLimited:
      description: Demasiadas peticiones
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Too many requests, please try again later."